+++ buildSrc/src/main/kotlin/io/getstream/chat/android/command/dag/ktlint/plugin/KtCommandExtesion.kt
@@ -1,6 +1,0 @@
-package io.getstream.chat.android.command.dag.ktlint.plugin
-
-open class KtCommandExtesion {
-    var changeModulesPath: String = "."
-    var outputPath: String = "kt-selected.sh"
-}