+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/attachments/factory/GiphyAttachmentFactory.kt
@@ -7,10 +7,14 @@ import io.getstream.chat.android.compose.ui.attachments.content.GiphyAttachmentC
 
 /**
  * An [AttachmentFactory] that validates and shows Giphy attachments using [GiphyAttachmentContent].
+ *
+ * Has no "preview content", given that this attachment only exists after being sent.
+ *
  */
 @Suppress("FunctionName")
 public fun GiphyAttachmentFactory(): AttachmentFactory = AttachmentFactory(
     canHandle = { attachments -> attachments.any { it.type == ModelType.attach_giphy } },
+    previewContent = @Composable { _, _, _ -> },
     content = @Composable { modifier, state ->
         GiphyAttachmentContent(
             modifier = modifier,