+++ stream-chat-android-core/src/main/java/io/getstream/chat/android/client/call/Call.kt
@@ -91,6 +91,10 @@ public fun <T : Any> Call<T>.doOnResult(scope: CoroutineScope, function: suspend
     DoOnResultCall(this, scope, function)
 
 @InternalStreamChatApi
+public fun <T: Any> Call<T>.withPrecondition(scope: CoroutineScope, precondition: suspend () -> Result<Unit>): Call<T> =
+    WithPreconditionCall(this, scope, precondition)
+
+@InternalStreamChatApi
 public fun Call<*>.toUnitCall(): Call<Unit> = map {}
 
 private val onSuccessStub: (Any) -> Unit = {}